Restructured reachability analysis
Split the modules into definitions, utilities and the general/free choice analyses
Rewrote imports accordingly
