+++ src/Language/FCL/Compile.hs
@@ -72,7 +72,7 @@ import qualified Language.FCL.Parser as Parser
 import qualified Language.FCL.Duplicate as Dupl
 import qualified Language.FCL.Effect as Effect
 import qualified Language.FCL.Typecheck as Typecheck
-import qualified Language.FCL.ReachabilityGraph as Reachability
+import qualified Language.FCL.Reachability.FreeChoice as Reachability
 import qualified Language.FCL.Undefinedness as Undef
 import Language.FCL.Warning (Warning(..))
 import Language.FCL.Utils ((?))
