+++ src/Language/FCL/Eval.hs
@@ -51,7 +51,7 @@ import Language.FCL.Storage
 import Language.FCL.Error as Error
 import Language.FCL.Prim (PrimOp(..))
 import Language.FCL.Address as Addr
-import Language.FCL.ReachabilityGraph (applyTransition)
+import Language.FCL.Reachability.Utils (applyTransition)
 import Language.FCL.Utils (panicImpossible)
 import Language.FCL.Asset as Asset
 import qualified Language.FCL.Delta as Delta
